added T5Chat component
